Unity updates & general improvements.
Update to 2018.4.1, update clipping.


Update to 2018.4.2


Update to 2019.1.6


New fullscreen toggle button.

- Create button to toggle fullscreen mode in the bottom left.
- Added ability to show cursor by pressing escape. The cursor is hidden again when there is movement input.
- Removed VisualStudio template from .gitignore.
- Added asset meta files.
- Switch to forced text mode.
- Updated to Unity 2019.1.7f1.

Update to 2019.1.8


Update to 2019.1.9


Reduce pickup count to be more reasonable.


Update to 2019.1.10


Update win text.


Update to 2019.1.11


Update to 2019.1.12


Update to 2019.2.0


Update to 2019.2.1


Update to 2019.2.2


Update to 2019.2.3


Update to 2019.2.6


Update to 2019.2.7


Update to 2019.2.8


Update to 2019.2.9


Update to 2019.2.10


Update to 2019.2.11


Update to 2019.2.12


Update to 2019.2.13


Update to 2019.3